CLdb_getPAMs.pl: CRISPR_array = 'no'

+++ blib/script/CLdb_delete.pl
@@ -68,29 +68,42 @@ write_selected_loci($loci_r) unless $verbose;
 ## spacer_pairwise_blast
 	# '*' = simple join on locus_id
 
+# listing tables to delete #
+my $tables_r = list_tables($dbh);
+
 # need to join; get unique IDs and delete on unique IDs #
 ## hclust tables ##
-delete_hclust($dbh, $loci_r, "spacers", "spacer_id", "spacer_hclust");
-delete_hclust($dbh, $loci_r, "DirectRepeats", "repeat_id", "directrepeat_hclust");
+delete_hclust($dbh, $loci_r, "spacers", "spacer_id", "spacer_hclust") 
+	if exists $tables_r->{"spacers_hclust"};
+delete_hclust($dbh, $loci_r, "DirectRepeats", "repeat_id", "directrepeat_hclust")
+	if exists $tables_r->{"directrepeat_hclust"};
 
 ## blast tables ##
-delete_spacer_pairwise_blast($dbh, $loci_r);
+delete_spacer_pairwise_blast($dbh, $loci_r) if exists $tables_r->{"spacer_pairwise_blast"};
 
 
 # deleting tables without needing to join; locus_id present #
-simple_delete($dbh, $loci_r, "spacers");
-simple_delete($dbh, $loci_r, "directrepeats");
-simple_delete($dbh, $loci_r, "directrepeatconsensus");
-simple_delete($dbh, $loci_r, "leaderseqs");
-simple_delete($dbh, $loci_r, "draft");
-simple_delete($dbh, $loci_r, "genes");
-simple_delete($dbh, $loci_r, "loci");
+simple_delete($dbh, $loci_r, "spacers") if exists $tables_r->{"spacers"};
+simple_delete($dbh, $loci_r, "directrepeats") if exists $tables_r->{"directrepeats"};
+simple_delete($dbh, $loci_r, "directrepeatconsensus") if exists $tables_r->{"directrepeatconsensus"};
+simple_delete($dbh, $loci_r, "leaderseqs") if exists $tables_r->{"leaderseqs"};
+simple_delete($dbh, $loci_r, "draft") if exists $tables_r->{"draft"};
+simple_delete($dbh, $loci_r, "genes") if exists $tables_r->{"genes"};
+simple_delete($dbh, $loci_r, "loci") if exists $tables_r->{"loci"};
 
 
 $dbh->commit();
 exit;
 
 ### Subroutines
+sub list_tables{
+	my $dbh = shift;
+	my $all = $dbh->selectall_hashref("SELECT tbl_name FROM sqlite_master", 'tbl_name');
+	map{delete $all->{$_}; tr/A-Z/a-z/; $all->{$_} = 1} keys %$all;
+		#print Dumper %$all; exit;
+	return $all;
+	}
+
 sub delete_spacer_pairwise_blast{
 	my ($dbh, $loci_r) = @_;
